database-service/requirements.txt
flask==2.0.1
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.6.1
werkzeug==2.0.3
boto3==1.34.34
email-validator==2.1.0 